Jupyter notebook for Diffusion informed pruning
-We use the diffusion distances to inform whether or not an edge should be kept in the original graph
-In particular we do 2 knn searches, the first on the Euclidean distances, we then build a affinity matrix and 
diffuse over that for some number of time steps. Then we do a knn search on the eigenspace. The philosophy is
that if a nn persists in Eigenspace then it has been walked on a lot and hence it is a 'true edge' in the Euclidean
graph. We prune all other edges
-The graph construction in Euclidean space uses gaussian weights to encourage very sparse inputs to the diffusion;
this is like taking the best from both worlds for small k in knn, i.e. the super sparse Euclidean graph has effective k very low,
 once propagated via diffusion, the walk heuristically takes this k^t (t for time steps) in the sense that is what is seen in the diffusion
-better put: propagating local information turns into global information which we can use the original graph
-also normalized the hyper parameters so that we don't have annoying numbers anymore